requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
discord.py==1.7.1
idna==3.1
macholib==1.14
multidict==5.1.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
typing-extensions==3.7.4.3
yarl==1.6.3
